+++ backend/UserServer/src/main/java/com/kl1verse/UserServer/domain/auth/repository/entity/Token.java
@@ -2,15 +2,12 @@ package com.kl1verse.UserServer.domain.auth.repository.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.kl1verse.UserServer.domain.user.repository.entity.User;
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -29,7 +26,7 @@ public class Token {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "token_id")
-    private Long id;
+    private Integer id;
 
 
     @Column(name = "refresh_token")
